+++ implementation/src/main/java/com/avioconsulting/mule/testing/muleinterfaces/wrappers/EventWrapper.groovy
@@ -34,4 +34,6 @@ interface EventWrapper {
                                          Map attributes,
                                          ConnectorInfo connectorInfo,
                                          boolean useRepeatableStream)
+
+    String getCorrelationId()
 }