+++ implementation/src/main/java/com/avioconsulting/mule/testing/muleinterfaces/wrappers/EventWrapperImpl.groovy
@@ -106,6 +106,11 @@ class EventWrapperImpl implements
     }
 
     @Override
+    String getCorrelationId() {
+        nativeEvent.correlationId
+    }
+
+    @Override
     MessageWrapper getMessage() {
         message
     }